Renamed components folder to use lower case
