+++ src/components/FireBase/Storage/fileManipulation.js
